@@ -11,10 +11,10 @@ description: Application which allows to bring more competition in your game.
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.5+1
+version: 1.0.6+2
 
 environment:
-  sdk: ">=2.5.2 <3.0.0"
+  sdk: ">=2.16.2 <3.0.0"
 
 dependencies:
   flutter:
@@ -22,9 +22,9 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  provider: ^4.0.2
-  hive: ^1.4.1+1
+  cupertino_icons: ^1.0.2
+  provider: ^6.0.2
+  hive: ^2.1.0
   hive_flutter: any
 
 dev_dependencies:
