@@ -6,10 +6,10 @@ import 'package:hive/hive.dart';
 
 class PlayersModel with ChangeNotifier {
   bool _boxInitialized = false;
-  String _activeSeason;
+  String? _activeSeason;
   bool isWithNew = false;
-  int indexOfEdit;
-  String _key;
+  int? indexOfEdit;
+  String? _key;
 
   updateActiveSeason(String newActiveSeason) {
     _boxInitialized = false;
@@ -25,7 +25,7 @@ class PlayersModel with ChangeNotifier {
     if (newActiveSeason != _activeSeason) {
       _activeSeason = newActiveSeason;
       _key = "players_$_activeSeason";
-      await Hive.openBox<Player>(_key);
+      await Hive.openBox<Player>(_key!);
     }
   }
 
@@ -33,7 +33,7 @@ class PlayersModel with ChangeNotifier {
     if (!_boxInitialized || _key == null) {
       return [];
     } else {
-      var box = Hive.box<Player>(_key);
+      var box = Hive.box<Player>(_key!);
       return box.values.toList();
     }
   }
@@ -77,7 +77,7 @@ class PlayersModel with ChangeNotifier {
       return;
     }
     isWithNew = false;
-    var box = Hive.box<Player>(_key);
+    var box = Hive.box<Player>(_key!);
     box.add(Player(name));
 
     notifyListeners();
