@@ -8,18 +8,18 @@ part of 'game.dart';
 
 class GameAdapter extends TypeAdapter<Game> {
   @override
-  final typeId = 1;
+  final int typeId = 1;
 
   @override
   Game read(BinaryReader reader) {
-    var numOfFields = reader.readByte();
-    var fields = <int, dynamic>{
-      for (var i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
+    final numOfFields = reader.readByte();
+    final fields = <int, dynamic>{
+      for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
     };
     return Game(
       fields[0] as int,
-      (fields[1] as List)?.cast<int>(),
-      (fields[2] as List)?.cast<int>(),
+      (fields[1] as List).cast<int>(),
+      (fields[2] as List).cast<int>(),
       fields[3] as int,
       fields[4] as int,
       fields[5] as int,
@@ -49,4 +49,14 @@ class GameAdapter extends TypeAdapter<Game> {
       ..writeByte(7)
       ..write(obj.teamWinByPenalty);
   }
+
+  @override
+  int get hashCode => typeId.hashCode;
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is GameAdapter &&
+          runtimeType == other.runtimeType &&
+          typeId == other.typeId;
 }
