@@ -6,8 +6,8 @@ import 'package:hive/hive.dart';
 
 class GamesModel with ChangeNotifier {
   bool _boxInitialized = false;
-  String _activeSeason;
-  String _key;
+  String _activeSeason = "";
+  String _key = "";
 
   updateActiveSeason(String newActiveSeason) {
     _boxInitialized = false;
