@@ -8,13 +8,13 @@ part of 'player.dart';
 
 class PlayerAdapter extends TypeAdapter<Player> {
   @override
-  final typeId = 0;
+  final int typeId = 0;
 
   @override
   Player read(BinaryReader reader) {
-    var numOfFields = reader.readByte();
-    var fields = <int, dynamic>{
-      for (var i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
+    final numOfFields = reader.readByte();
+    final fields = <int, dynamic>{
+      for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
     };
     return Player(
       fields[0] as String,
@@ -30,4 +30,14 @@ class PlayerAdapter extends TypeAdapter<Player> {
       ..writeByte(1)
       ..write(obj.inClub);
   }
+
+  @override
+  int get hashCode => typeId.hashCode;
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is PlayerAdapter &&
+          runtimeType == other.runtimeType &&
+          typeId == other.typeId;
 }
